LayoutTests/accessibility/ios-simulator/element-text-range-for-text-control.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script>
var successfullyParsed = false;
</script>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input id="input" type='text'>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This test ensures that elementTextRange is valid for text controls");

    if (window.accessibilityController) {

        var input = accessibilityController.accessibleElementById("input");
        shouldBeTrue("input.elementTextPosition != -1");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
