LayoutTests/accessibility/ios-simulator/element-text-range-for-text-control-expected.txt


This test ensures that elementTextRange is valid for text controls

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.elementTextPosition != -1 is true
PASS successfullyParsed is true

TEST COMPLETE
